Setup dependency injection, use dependency injection in queue handlers and behaviour
