+++ MtaServer.Server/Behaviour/DefaultChatBehaviour.cs
@@ -4,12 +4,13 @@ using System.Drawing;
 using System.Text;
 using MtaServer.Packets.Definitions.Commands;
 using MtaServer.Server.Elements;
+using MtaServer.Server.Repositories;
 
 namespace MtaServer.Server.Behaviour
 {
     public class DefaultChatBehaviour
     {
-        public DefaultChatBehaviour(MtaServer server)
+        public DefaultChatBehaviour(IElementRepository elementRepository)
         {
             Player.OnJoin += (player) =>
             {
@@ -17,8 +18,8 @@ namespace MtaServer.Server.Behaviour
                 {
                     if(command == "say")
                     {
-                        var packet = new ChatEchoPacket(server.Root.Id, player.Name + ": " + string.Join(' ', arguments), Color.White);
-                        foreach (var _player in server.ElementRepository.GetByType<Player>(ElementType.Player))
+                        var packet = new ChatEchoPacket(player.Id, player.Name + ": " + string.Join(' ', arguments), Color.White);
+                        foreach (var _player in elementRepository.GetByType<Player>(ElementType.Player))
                         {
                             _player.Client.SendPacket(packet);
                         }
