+++ MtaServer.Server/Elements/Enums/MarkerType.cs
