+++ MtaServer.Server/Elements/Enums/PickupType.cs
