+++ MtaServer.Server/Elements/Enums/WeaponEnums.cs
