+++ MtaServer.Server/Elements/RootElement.cs
@@ -1,0 +1,10 @@
+﻿using System;
+using System.Numerics;
+
+namespace MtaServer.Server.Elements
+{
+    public class RootElement: Element
+    {
+        public override ElementType ElementType => ElementType.Root;
+    }
+}
