+++ MtaServer.Server/MtaServer.Server.csproj
@@ -8,6 +8,7 @@
 
   <ItemGroup>
     <PackageReference Include="KdTree" Version="1.4.1" />
+    <PackageReference Include="Microsoft.Extensions.DependencyInjection" Version="5.0.0-preview.6.20305.6" />
   </ItemGroup>
 
   <ItemGroup>
