+++ MtaServer.Server/PacketHandling/QueueHandlers/WorkerBasedQueueHandler.cs
@@ -7,12 +7,10 @@ namespace MtaServer.Server.PacketHandling.QueueHandlers
 {
     public abstract class WorkerBasedQueueHandler : BaseQueueHandler
     {
-        protected readonly MtaServer server;
         protected readonly int sleepInterval;
 
-        public WorkerBasedQueueHandler(MtaServer server, int sleepInterval = 10, int workerCount = 1) : base()
+        public WorkerBasedQueueHandler(int sleepInterval = 10, int workerCount = 1) : base()
         {
-            this.server = server;
             this.sleepInterval = sleepInterval;
 
             for (int i = 0; i < workerCount; i++)
